Refactor GameViewResolutionAttribute
