+++ Runtime/Attributes/GameViewResolutionAttribute.cs
@@ -5,11 +5,7 @@ using System;
 using NUnit.Framework;
 using NUnit.Framework.Interfaces;
 using NUnit.Framework.Internal;
-using TestHelper.RuntimeInternals.Wrappers.UnityEditor;
-using UnityEngine;
-#if UNITY_EDITOR
-using UnityEditor;
-#endif
+using TestHelper.RuntimeInternals;
 
 namespace TestHelper.Attributes
 {
@@ -48,63 +44,7 @@ namespace TestHelper.Attributes
         /// <inheritdoc />
         public void ApplyToContext(ITestExecutionContext context)
         {
-#if UNITY_2022_2_OR_NEWER
-            SetResolutionUsingPlayModeWindow();
-#else
-            SetResolution();
-#endif
-        }
-
-        // ReSharper disable once UnusedMember.Local
-        private void SetResolutionUsingPlayModeWindow()
-        {
-#if UNITY_EDITOR && UNITY_2022_2_OR_NEWER
-            PlayModeWindow.SetViewType(PlayModeWindow.PlayModeViewTypes.GameView);
-            PlayModeWindow.SetCustomRenderingResolution(_width, _height, _name);
-#endif
-        }
-
-        // ReSharper disable once UnusedMember.Local
-        private void SetResolution()
-        {
-#if UNITY_EDITOR
-            var gameViewSizes = GameViewSizesWrapper.CreateInstance();
-            if (gameViewSizes == null)
-            {
-                Debug.LogError("GameViewSizes instance creation failed.");
-                return;
-            }
-
-            var gameViewSizeGroup = gameViewSizes.CurrentGroup();
-            if (gameViewSizeGroup == null)
-            {
-                Debug.LogError("GameViewSizeGroup instance creation failed.");
-                return;
-            }
-
-            var gameViewSize = GameViewSizeWrapper.CreateInstance((int)_width, (int)_height, _name);
-            if (gameViewSize == null)
-            {
-                Debug.LogError("GameViewSize instance creation failed.");
-                return;
-            }
-
-            var index = gameViewSizeGroup.IndexOf(gameViewSize);
-            if (index == -1)
-            {
-                gameViewSizeGroup.AddCustomSize(gameViewSize);
-                index = gameViewSizeGroup.IndexOf(gameViewSize);
-            }
-
-            var gameView = GameViewWrapper.GetWindow();
-            if (gameView == null)
-            {
-                Debug.LogError("GameView instance creation failed.");
-                return;
-            }
-
-            gameView.SelectedSizeIndex(index);
-#endif
+            GameViewControlHelper.SetResolution(_width, _height, _name);
         }
     }
 }
