+++ RuntimeInternals/GameViewControlHelper.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Reflection;
+using TestHelper.RuntimeInternals.Wrappers.UnityEditor;
 using UnityEditor;
 using UnityEngine;
 
@@ -31,6 +32,73 @@ namespace TestHelper.RuntimeInternals
             }
 
             EditorWindow.GetWindow(s_gameView, false, null, true);
+#endif
+        }
+
+        /// <summary>
+        /// Set <c>GameView</c> resolution.
+        /// </summary>
+        /// <param name="width">GameView width [px]</param>
+        /// <param name="height">GameView height [px]</param>
+        /// <param name="name">GameViewSize name</param>
+        public static void SetResolution(uint width, uint height, string name)
+        {
+#if UNITY_2022_2_OR_NEWER
+            SetResolutionUsingPlayModeWindow(width, height, name);
+#else
+            SetResolutionUsingReflection(width, height, name);
+#endif
+        }
+
+        // ReSharper disable once UnusedMember.Local
+        private static void SetResolutionUsingPlayModeWindow(uint width, uint height, string name)
+        {
+#if UNITY_EDITOR && UNITY_2022_2_OR_NEWER
+            PlayModeWindow.SetViewType(PlayModeWindow.PlayModeViewTypes.GameView);
+            PlayModeWindow.SetCustomRenderingResolution(width, height, name);
+#endif
+        }
+
+        // ReSharper disable once UnusedMember.Local
+        private static void SetResolutionUsingReflection(uint width, uint height, string name)
+        {
+#if UNITY_EDITOR
+            var gameViewSizes = GameViewSizesWrapper.CreateInstance();
+            if (gameViewSizes == null)
+            {
+                Debug.LogError("GameViewSizes instance creation failed.");
+                return;
+            }
+
+            var gameViewSizeGroup = gameViewSizes.CurrentGroup();
+            if (gameViewSizeGroup == null)
+            {
+                Debug.LogError("GameViewSizeGroup instance creation failed.");
+                return;
+            }
+
+            var gameViewSize = GameViewSizeWrapper.CreateInstance((int)width, (int)height, name);
+            if (gameViewSize == null)
+            {
+                Debug.LogError("GameViewSize instance creation failed.");
+                return;
+            }
+
+            var index = gameViewSizeGroup.IndexOf(gameViewSize);
+            if (index == -1)
+            {
+                gameViewSizeGroup.AddCustomSize(gameViewSize);
+                index = gameViewSizeGroup.IndexOf(gameViewSize);
+            }
+
+            var gameView = GameViewWrapper.GetWindow();
+            if (gameView == null)
+            {
+                Debug.LogError("GameView instance creation failed.");
+                return;
+            }
+
+            gameView.SelectedSizeIndex(index);
 #endif
         }
     }
